[RTL] Fix MSVC warning about signed/unsigned mismatch. In this case it was not a bug, because the compiler implicitly casts to unsigned, when signed and unsigned are compared, so (-1 < 0x23U) is false.
svn path=/trunk/; revision=53163

+++ reactos/lib/rtl/error.c
@@ -56,8 +56,8 @@ ULONG WINAPI RtlNtStatusToDosErrorNoTeb( NTSTATUS status )
 
     while (table->start)
     {
-        if (status < table->start) break;
-        if (status < table->end)
+        if ((ULONG)status < table->start) break;
+        if ((ULONG)status < table->end)
         {
             DWORD ret = table->table[status - table->start];
             /* unknown entries are 0 */
